Game: provide player count and player summary properties
We only need the full info of the player whose tab we are actively
showing. For all other players, we only need just enough info to display
their tab.

Queries before this commit:

* one to check whether there are any players
* one (full info, unnecessary) to get the ID of the first player so as
  to show their tab
* three times to get the count:
  * once to check if it's non-zero to decide whether to show tabs
  * twice for the discard N cards from deck button
* once to show the full info of the current tab's player
* once (full info, unnecessary) to show the tab list

Queries after this commit:

* once to get the count of players
* once to get the ID of the first player
* once to show the full info of the current tab's player
* once (summary) to show the tab list

You'd think the count and summary queries could be combined, but note
that the tabs are a different HTTP request from the main game page and
therefore can't share the same instance of Game.

Whether this is possible to resolve is for future investigation; it's
outside the scope of this commit.